[storage] background downloader queue is unified
