+++ storage/CMakeLists.txt
@@ -51,7 +51,12 @@ set(
   storage_helpers.hpp
 )
 
-if (${PLATFORM_IPHONE})
+if (${PLATFORM_IPHONE} OR ${PLATFORM_ANDROID})
+  append(
+    SRC
+    background_downloading/downloader_queue.hpp
+  )
+elseif (${PLATFORM_IPHONE})
   append(
     SRC
     background_downloading/downloader_adapter_ios.h
